+++ src/FIMS.cpp
@@ -1,7 +1,7 @@
 
 //#include "../inst/include/interface/interface.hpp"
 
-#include "../inst/include/interface/rcpp_interface.hpp"
+#include "../inst/include/interface/rcpp/rcpp_interface.hpp"
 #include "init.hpp"
